Refactor to use public type definition file Fixes #45
