+++ package.json
@@ -51,6 +51,7 @@
     "@types/express": "^4.17.6",
     "@types/node": "^13.1.2",
     "@types/node-fetch": "^2.5.7",
+    "@types/node-red": "^0.20.1",
     "@types/request-promise-native": "^1.0.17",
     "@typescript-eslint/eslint-plugin": "^3.5.0",
     "@typescript-eslint/parser": "^3.5.0",
