+++ src/types/IWledNode.ts
@@ -2,9 +2,12 @@
  *  Copyright (c) Neil Enns. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
+import WledDevice from "../WledDevice";
+import * as Red from "node-red";
+import IWledNodeProperties from "./IWledNodeProperties";
 
-import INodeConfig from "./INodeConfig";
-
-export default interface INodeRedMessage {
-  payload: INodeConfig;
+export default interface IWledNode extends Red.Node {
+  wled: WledDevice;
+  config: IWledNodeProperties;
+  solidTimer: NodeJS.Timeout;
 }
